refactor: HeaderSearch component, fix: search modal close on search
