+++ src/components/header/HeaderComponent.vue
@@ -9,6 +9,7 @@ import {
 } from 'vue'
 
 import type { DsfrHeaderMenuLinkProps } from './DsfrHeaderMenuLink.vue'
+import HeaderSearch from './HeaderSearch.vue'
 
 type DsfrHeaderProps = {
   serviceTitle?: string
@@ -45,10 +46,9 @@ const props = withDefaults(defineProps<DsfrHeaderProps>(), {
   searchLabel: 'Recherche',
   quickLinksAriaLabel: 'Menu secondaire',
   showBadge: false,
-  badgeText: 'BETA'
+  badgeText: 'BETA',
+  badgeStyle: ''
 })
-
-const modelText = ref()
 
 const onKeyDown = (e: KeyboardEvent) => {
   if (e.key === 'Escape') {
@@ -65,10 +65,12 @@ onUnmounted(() => {
 const searchModalOpened = ref(false)
 const headerModalOpened = ref(false)
 
-const hideModal = () => {
+const hideModal = (focusButtonMenu: boolean = true) => {
   headerModalOpened.value = false
   searchModalOpened.value = false
-  document.getElementById('button-menu')?.focus()
+  if (focusButtonMenu) {
+    document.getElementById('button-menu')?.focus()
+  }
 }
 const showMenu = () => {
   headerModalOpened.value = true
@@ -88,17 +90,6 @@ const isWithSlotOperator = computed(
 const isWithSlotNav = computed(() => Boolean(slots.mainnav))
 
 const badgeCss = 'fr-badge fr-badge--sm fr-badge--' + props.badgeStyle
-
-// eslint-disable-next-line func-call-spacing
-const emit = defineEmits<{
-  (e: 'update:modelValue', payload: string): void
-  (e: 'search', payload: string): void
-}>()
-
-function search(payload: string) {
-  emit('search', payload)
-  modelText.value = ''
-}
 </script>
 
 <template>
@@ -209,13 +200,9 @@ function search(payload: string) {
               </nav>
             </div>
             <div v-if="showSearch" class="fr-header__search fr-modal">
-              <DsfrSearchBar
-                :label="searchLabel"
-                v-model="modelText"
+              <HeaderSearch
+                :search-label="searchLabel"
                 :placeholder="placeholder"
-                style="justify-content: flex-end"
-                @update:model-value="$emit('update:modelValue', $event)"
-                @search="search(modelText)"
               />
             </div>
           </div>
@@ -239,12 +226,11 @@ function search(payload: string) {
             >
               Fermer
             </button>
-            <DsfrSearchBar
+            <HeaderSearch
               v-if="searchModalOpened"
-              v-model="modelText"
+              :search-label="searchLabel"
               :placeholder="placeholder"
-              @update:model-value="$emit('update:modelValue', $event)"
-              @search="search(modelText)"
+              @search="hideModal(false)"
             />
           </div>
         </div>
